requirements.txt
openai 
fastapi[all] 
uvicorn[standard] 
openai-whisper 
jiwer
pydub